- Remove spawning on RandomSpawners.
I don't feel like this is a good way to do this at all honestly as it gives the potential for stuff that doesn't need to be replaced getting replaces but it's what seems to work with this handler so unless I fix it I'm not gonna worry about it.
Some other time I'll add more things it can possibly spawn on but that's for later as it'd require adding even more lines to this.